UINT??をuint??_tに統一しました

+++ Engine/Rendering/CabotEngine/Graphics/Texture2D.cpp
@@ -16,12 +16,12 @@ Texture2D::Texture2D(const aiTexture *ai_texture)
 {
     unsigned char *pixels;
     int width = 0, height = 0, channels = 0;
-
+    
     if (ai_texture->mHeight == 0)
     {
         pixels = stbi_load_from_memory(
             reinterpret_cast<const unsigned char *>(ai_texture->pcData),
-            ai_texture->mWidth,
+            static_cast<int>(ai_texture->mWidth),
             &width,
             &height,
             &channels,
@@ -31,9 +31,8 @@ Texture2D::Texture2D(const aiTexture *ai_texture)
     }
     else
     {
-        // RGBAのRAWデータ
-        width = ai_texture->mWidth;
-        height = ai_texture->mHeight;
+        width = static_cast<int>(ai_texture->mWidth);
+        height = static_cast<int>(ai_texture->mHeight);
         pixels = reinterpret_cast<unsigned char *>(ai_texture->pcData);
     }
     m_width_ = width;
